Month segmentation seems to be correct

+++ Classes/Data/Month.h
@@ -10,16 +10,16 @@
 
 @interface Month : NSObject {
     NSString *strDate;
-    float min_ts;
-    float max_ts;
+    double min_ts;
+    double max_ts;
     int count;
 }
 
 @property (readwrite,retain)  NSString *strDate;
-@property (readwrite)  float min_ts;
-@property (readwrite)  float max_ts;
+@property (readwrite)  double min_ts;
+@property (readwrite)  double max_ts;
 @property (readwrite)  int count;
 
-- (id) initWithData:(NSString*)str min_ts:(float)mints max_ts:(float)maxts count:(int)c ;
+- (id) initWithData:(NSString*)str min_ts:(double)mints max_ts:(double)maxts count:(int)c ;
 
 @end
